uschess: cache requests
The latency of responses from most of the uschess.org endpoints is
quite high, often exceeding 2s. This presents a challenge for the
discordbot which only has 3s total to reply. A lambda cold start plus
seqential latencies from multiple requests to these endpoints can
often exceed this budget. To mitigate, this commit introduces a cache
for this data as it changes very slowly or not at all.