discordbot: add crosstable sub command
This commit adds a crosstable sub command to retrieve the crosstables
for a completed tournament. Currently the command only supports
tournaments which have recorded their results with USCF. In the future
we should be able to support crosstables for currently running
tournaments. This is enabled by the bcc API's recent export of MsaEventId
in event detail.